[feat] : 회원정보에 좋아요한 게시글과 댓글 갯수 추가

+++ src/main/java/com/sparta/outsourcing/like/dto/LikeCountResponseDto.java
@@ -6,6 +6,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class LikeCountResponseDto {
-
-    private int count;
+    private Long count;
 }
